+++ src/main/java/org/esfe/repositorios/ICompraRepository.java
@@ -3,6 +3,6 @@ package org.esfe.repositorios;
 import org.esfe.modelos.Compra;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICompraRepository extends JpaRepository<Compra, Integer> {
+public interface ICompraRepository extends JpaRepository<Compra, Long> {
 
 }
