+++ src/main/java/org/esfe/repositorios/IDetalleCompraRespository.java
@@ -3,6 +3,6 @@ package org.esfe.repositorios;
 import org.esfe.modelos.DetalleCompra;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IDetalleCompraRespository extends JpaRepository <DetalleCompra, Integer> {
+public interface IDetalleCompraRespository extends JpaRepository <DetalleCompra, Long> {
 
 }
