+++ src/main/java/org/esfe/repositorios/IDetalleVentaRepository.java
@@ -3,5 +3,5 @@ package org.esfe.repositorios;
 import org.esfe.modelos.DetalleVenta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IDetalleVentaRepository extends JpaRepository<DetalleVenta, Integer> {
+public interface IDetalleVentaRepository extends JpaRepository<DetalleVenta, Long> {
 }
