+++ src/main/java/org/esfe/repositorios/IInventarioRepository.java
@@ -3,5 +3,5 @@ package org.esfe.repositorios;
 import org.esfe.modelos.Inventario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IInventarioRepository extends JpaRepository<Inventario, Integer> {
+public interface IInventarioRepository extends JpaRepository<Inventario, Long> {
 }
