+++ src/main/java/org/esfe/repositorios/IVentaRepository.java
@@ -3,5 +3,5 @@ package org.esfe.repositorios;
 import org.esfe.modelos.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IVentaRepository extends JpaRepository<Venta, Integer> {
+public interface IVentaRepository extends JpaRepository<Venta, Long> {
 }
